+++ .github/workflows/deploy.yml
@@ -19,7 +19,7 @@ jobs:
         logs: https://github.com/npm/action-deploy/actions?query=workflow%3Adeploy
         environment_url: https://npmjs.com
         job_status: ${{job.status}}
-        slack_token: ${{secrets.SLACK_NPM_ROBOT}}
+        slack_token: ${{secrets.NPM_ROBOT_SLACK_TOKEN}}
         slack_channel: npm-ops
 
     # add here an actual deployment
